Create task 4
I have tried running the code on PyCHarm 2020.3.3. In order to even tried running it, I had to delet the comment and the beggining of the code, but I have encountered an issue. I have updated the operating system and the  PyCHarm's default interactive shell is now zsh. I had to switch the shell from bin to zsh. Wan's local machine read the code without an issue and Siergiey's PyCharm (the same version) is announcing that he has to dowload libraries in order to run the code.